Fix StyleCop warnings SA1402–SA1414 and SA1629
AB#9817
