+++ src/NHSD.BuyingCatalogue.Ordering.Domain/Common/ValueObject.cs
@@ -40,7 +40,7 @@ namespace NHSD.BuyingCatalogue.Ordering.Domain.Common
                 {
                     unchecked
                     {
-                        return current * 23 + (obj?.GetHashCode() ?? 0);
+                        return (current * 23) + (obj?.GetHashCode() ?? 0);
                     }
                 });
         }
